Move payloads persistence into the state package.
